Renamed Tags to TagTypes
because it only has types
and i need to create a list of Tags
